Fixed transforms and the second pass of training
Dataloader now handles transforms for image and mask correctly by stacking and applying the transform to the combined data.  Can pass a joint_transform, image_transform, and mask_transform.

I found an issue with the second pass of training.  It was starting with the pretrained imagenet model.  It should be picking up with the model trained in the first pass.  So now it loads those models from storage before training in the second pass.